implement DVC in data ingestion process to handle data versioning
