+++ .dvc/config
@@ -1,0 +1,4 @@
+[core]
+    remote = storage
+['remote "storage"']
+    url = s3://ml-feature-store-029552764749/credit_fraud
