--- artifacts/data.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0e1d0604d66a03f78fcd3970dce63262a3a7c32d67a8b505b6ef264c4e969200
-size 189945256
